Remove power analysis.
The role of power analysis was to try and estimate what a good
sample size would be in order to be able to detect a difference
of > or < 0, assuming the difference was some small amount.

Now that we have automatic sampling, though, this way of thinking
about the sample size isn't really useful anymore.

I think it could still be useful for estimating how many samples
are going to be needed during automatic sampling, which could be
used for giving a nice time estimate, but for now I'm just going
to remove it.